requirements (1).txt
setuptools
wheel
aiohttp
python-dotenv
python-telegram-bot
sumy
nltk
tenacity
python-dateutil